services/core_requirements.txt
fastapi
uvicorn
aiohttp
pyyaml
schedule
python-telegram-bot
pydantic-settings
openai
python-dotenv